--- .idea/shelf/Uncommitted_changes_before_Checkout_at_2024-06-03__11_38__Changes_.xml
@@ -1,4 +1,0 @@
-<changelist name="Uncommitted_changes_before_Checkout_at_2024-06-03_오전_11_38_[Changes]" date="1717382305668" recycled="true" deleted="true">
-  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_2024-06-03_오전_11_38_[Changes]/shelved.patch" />
-  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 2024-06-03 오전 11:38 [Changes]" />
-</changelist>
