+++ app/src/main/AndroidManifest.xml
@@ -36,8 +36,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".consumption.presentation.ConsumptionActivity"/>
-        <activity android:name=".consumption.presentation.CalculationActivity"/>
+        <activity android:name=".feature.consumption.presentation.ConsumptionActivity"/>
+        <activity android:name=".feature.consumption.presentation.CalculationActivity"/>
     </application>
 
 </manifest>
