--- app/src/main/java/com/example/donotlate/consumption/domain/usecase/GetFinishedConsumptionUseCase.kt
@@ -1,7 +1,0 @@
-package com.example.donotlate.consumption.domain.usecase
-
-import com.example.donotlate.consumption.domain.repository.ConsumptionRepository
-
-class GetFinishedConsumptionUseCase(private val repository: ConsumptionRepository) {
-    operator fun invoke() = repository.getFinishedConsumption()
-}
