--- app/src/main/java/com/example/donotlate/consumption/domain/usecase/GetUnfinishedConsumptionUseCase.kt
@@ -1,7 +1,0 @@
-package com.example.donotlate.consumption.domain.usecase
-
-import com.example.donotlate.consumption.domain.repository.ConsumptionRepository
-
-class GetUnfinishedConsumptionUseCase(private val repository: ConsumptionRepository) {
-    operator fun invoke() = repository.getUnfinishedConsumption()
-}
