+++ app/src/main/java/com/example/donotlate/feature/consumption/domain/usecase/GetConsumptionByCategoryUseCase.kt
@@ -1,0 +1,9 @@
+package com.example.donotlate.feature.consumption.domain.usecase
+
+import com.example.donotlate.feature.consumption.domain.repository.ConsumptionRepository
+
+class GetConsumptionByCategoryUseCase
+constructor(private val repository: ConsumptionRepository) {
+
+    suspend operator fun invoke(category: String) = repository.getConsumptionByCategory(category)
+}
