+++ app/src/main/java/com/example/donotlate/feature/consumption/domain/usecase/GetConsumptionByIdUseCase.kt
@@ -1,0 +1,8 @@
+package com.example.donotlate.feature.consumption.domain.usecase
+
+import com.example.donotlate.feature.consumption.domain.repository.ConsumptionRepository
+
+class GetConsumptionByIdUseCase
+constructor(private val repository: ConsumptionRepository) {
+    suspend operator fun invoke(historyId: String) = repository.getConsumptionById(historyId)
+}
