+++ app/src/main/java/com/example/donotlate/feature/consumption/domain/usecase/GetFinishedConsumptionUseCase.kt
@@ -1,0 +1,8 @@
+package com.example.donotlate.feature.consumption.domain.usecase
+
+import com.example.donotlate.feature.consumption.domain.repository.ConsumptionRepository
+
+class GetFinishedConsumptionUseCase
+constructor(private val repository: ConsumptionRepository) {
+    operator fun invoke() = repository.getRecentFinishedConsumption()
+}
