+++ app/src/main/java/com/example/donotlate/feature/consumption/domain/usecase/GetTotalPriceUseCase.kt
@@ -1,0 +1,8 @@
+package com.example.donotlate.feature.consumption.domain.usecase
+
+import com.example.donotlate.feature.consumption.domain.repository.ConsumptionRepository
+
+class GetTotalPriceUseCase
+constructor(private val repository: ConsumptionRepository){
+    operator fun invoke() = repository.getTotalPrice()
+}
