+++ app/src/main/java/com/example/donotlate/feature/room/presentation/dialog/BackFragmentDialog.kt
@@ -11,8 +11,8 @@ import androidx.fragment.app.DialogFragment
 import com.example.donotlate.MainActivity
 import com.example.donotlate.core.presentation.MainFragment
 import com.example.donotlate.databinding.BackDialogBinding
-import com.example.donotlate.feature.room.presentation.RoomStartFragment
-import com.example.donotlate.feature.room.presentation.ViewPagerFragment
+import com.example.donotlate.feature.room.presentation.start.RoomStartFragment
+import com.example.donotlate.feature.room.presentation.main.ViewPagerFragment
 
 class BackFragmentDialog : DialogFragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
